+++ frontend/src/App.jsx
@@ -9,7 +9,7 @@ import VisitForm from "./pages/VisitForm";
 import Navbar from "./components/Navbar";
 import PropTypes from "prop-types";
 import BrandForm from "./pages/BrandForm";
-import Reports from "./pages/Reports";
+import Reports from "./pages/ReportsForm";
 
 const PrivateRoute = ({ children }) => {
     const authContext = useContext(AuthContext);
